password: set keyring password enterprise accounts too
At the moment enteprise accounts end up with a keyring password
of "gis" because they return early from the password page, before
the keyring password updating code has an opportunity to run.

This commit fixes the code to update the keyring password for
enterprise accounts before returning on the password page.